requirements.txt
llama-cpp-python
duckduckgo-search
colorama
requests
beautifulsoup4
trafilatura
readchar
keyboard
windows-curses; sys_platform == 'win32'
tqdm
urllib3
openai>=1.0.0
anthropic>=0.7.0
PyYAML>=6.0
httpx>=0.24.0
aiofiles>=23.0.0
